+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # renv 0.14.0 (UNRELEASED)
 
+* `renv::install()` now better respects the `Remotes` field in a project
+  `DESCRIPTION` file, if available. (#670)
+
 * `RENV_DOWNLOAD_METHOD` is now treated as an alias for `RENV_DOWNLOAD_FILE_METHOD`.
 
 * Fixed an issue where `renv` would fail to load if the `~/.Rprofile` existed